pyproject.toml
[project]
name = "template"
version = "0.1.0"
readme = "README.md"
description = ""
authors = [
    { name = "MSFT" }
]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "dotenv>=0.9.9",
    "openai>=1.103.0",
    "azure-ai-projects>=1.0.0",
    "azure-identity>=1.24.0",
    "azure-cosmos==4.9.0",
    "opentelemetry-api>=1.19.0",
    "opentelemetry-sdk>=1.19.0",
    "opentelemetry-exporter-otlp-proto-http>=1.19.0",
    "azure-monitor-opentelemetry>=1.1.0",
    "tenacity>=8.2.2",
    "PyYAML>=6.0",
    "jinja2>=3.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.1",
    "pre-commit>=3.3.3"
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.ruff]
exclude = [
    ".git",
    ".pytest_cache",
    ".ruff_cache",
    ".vscode",
    "__pycache__",
    ".venv",
]

[tool.ruff.lint.pydocstyle]
convention = "google"